requirements.txt
cleo
gmusicapi
soundcloud
pyaml
cachy
pexpect
backpack
simplejson
